Fix two lines prompts in serial session
For windows guest (tested with Win2016 and Win2008), two line prompts
at same time in serial session.

as,
2018-11-13 00:42:12: C:\Windows\system32>
2018-11-13 00:42:12: C:\Windows\system32>
2018-11-13 00:42:13: C:\Windows\system32>
2018-11-13 00:42:13: tzutil /g
2018-11-13 00:42:13: SA Pacific Standard Time
2018-11-13 00:42:13: C:\Windows\system32>
2018-11-13 00:42:13: C:\Windows\system32>
2018-11-13 00:42:13: tzutil /l

We handle only one prompt in framework, the extra prompt will be
included in cmd result by mistake.
This is a bug for serialport communication tool, make a workaround
here to wait fix this bug.

Signed-off-by: Sitong Liu <siliu@redhat.com>