virttest/utils_test: use correct python on run_avocado()
The run_avocado() function attempt to install Avocado framework
and plugins in case they aren't present in the guest. However,
currently it assumes the guest has python2 installed, which
isn't always true. So this change the function to check what
python version is installed on guest (python3 is preferred), then
use the correct binary.

It was introduced the find_python() function in the library
that should be used to check and return the python binary
installed (None otherwise) in the guest.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>