virttest/utils_test: fixes to virtio serial file transfer
The run_virtio_serial_file_transfer() function runs python scripts
on (usually) both guest and host to test serial communication. It
assumes Python 2 is installed on both side, which is not always
true. So this change ensure it uses the python binary (python3 preferred)
available. It was added the find_host_python() function which is
used to find the python binary on host.

This change also contain some misc fixes:
a) Ensure proper comparisons of byte type with Python 3.
b) Ensure the working directory is created on the guest.
c) Fix the calculation of received bytes by the guest script.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>