virttests.qemu_devices: Stop VM to query for hmp/qmp commands
We use "live" qemu instance to query for supported QMP/human-monitor
commands, but the machine is not given any image. On most architectures
this only results to busy loop and unnecessary utilization for a short
period of time, but other architectures (like s390x) actually finishes
as soon as it finds no bootable image, which sometimes is before it
reports the commands. Let's use the '-S' to start in stopped state.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>