Makefile: get the version from setup.py
This has the downside of making things go through another layer, but
at the same time, it abstracts the implementation currently used for
versioning.  If we ever change the method for setting the version
number, there will be no update needed here.

Most importantly, this also syncs the behavior with Avocado.

Signed-off-by: Cleber Rosa <crosa@redhat.com>