Makefile: add a develop target
This gives a consistent experience if an Avocado-VT developer treats
Avocado itself as a library that is already available and installed
on its system.

This fulfills the use case where Avocado is installed by a distro
package and Avocado-VT is being developed on top of it.

It also syncs another target and experience from Avocado itself.

Signed-off-by: Cleber Rosa <crosa@redhat.com>