COPR SCM/"make srpm" support
This adds a Makefile that adds support for the Fedora COPR service to
build packages for Avocado-VT, right out of a Git repo.  To do so,
create your project, and in "Builds" tab, select "New Build".

On "1. Select source type" section, select "SCM".  On section
"2. Provide the source", under "Clone url" add the URL of your repo.
For instance, this has been tested with:

  https://github.com/clebergnu/avocado-vt

Under "Committish", enter a reference to a given commit.  For instance, this
has been tested with the name of this branch: "copr_makefile".

Under "3. How to build SRPM from the source", select "make srpm".
Finally, select the appropriate Chroots.  It's better not to have
"Enable internet access during this build" option checked.

Finally, it's advisable to not attempt to run this Makefile locally,
as it's designed to be run on the environment that CORP provides, that
is, a mock based chroot with super user privileges.

Signed-off-by: Cleber Rosa <crosa@redhat.com>