qemu_vm: Correct the check in `VM.mig_finished()`
Now we have many migration statuses other than `*active`, so we
could not make the judgment of a finished migration by verifying
the migration status is not `*active`. Hence, we should let it
return `True` only if the current status is in certain ones which
are `none`, `completed`, `failed` and `cancelled`.

Signed-off-by: Xu Han <xuhan@redhat.com>