virttest.qemu_monitor: Fix HMP get_migrate_parameter
This function hardcoded '"parameter"' instead of a variable "parameter"
and even with that it does not work properly with the current qemu,
because the output contains ":". Let's iterate through the returned
lines and yield the first match.

I'm not sure whether the format changed, but since it is broken for a
while I guess we don't need to double-check the backward compatibility.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>