qemu_vm: Delete all guest ports correctly when destroying guest
The current method sometimes cannot delete all ports of guest
accurately. Such as openvswitch backend, ovs takes some time to process
ports to be deleted. So we should not check the port status via 'ip
link'(get_net_if), we need to check if it exists in the bridge to decide
whether to delete it.

Signed-off-by: Yihuang Yu <yihyu@redhat.com>