virttest.qemu_vm: Protect check_go_down method from migration
The QMP check whether VM already rebooted fails in case the VM migrates,
because it keeps using the old (and dead) source VM's monitor. Let's
make sure we always use the current VM's monitor.

Using the new monitor is not actually sufficient as this situation might
happen:

1. vm starts migration
2. vm reboots, RESET even is signalled in src QMP
3. src vm finishes migration and dies
4. we check for events using src monitor and get MonitorSocketError
5. we check again now using dst vm QMP, but it does not contain the
   RESET event as it was already emitted in src QMP.

so false-negatives are still possible, but less likely. Let's log a
warning in such case. This warning would be logged every-time this
happens, but the probability of hitting this is minimal so it doesn't
seem worth logging it just once.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>