virttest.qemu_vm: Improve migration postprocess on failure
In case of post-copy migration the VM ends-up in "postcopy-paused"
state, that is a special case of the "paused" state. Our finally block
in migration function resumes the VM in such case which results in
exception as it is not possible to resume such VM.

This patch ignores monitor exceptions for the "resume" operation. This
should be fairly safe as different status would raise a different
exception (still we can fail to send "cont" for a different reason, but
then we should fail to interact with the VM and detect the failure
later)

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>