Introduce a new MAC generation algorithm
As the current algorithm randomly picks only one byte to generate
a MAC address, there are unignorable MAC collisions during multiple
job instances running in the same LAN simultaneously.

This patch introduces a new generation algorithm which is based on
UUID1 combined with SHA256 and expands the random field size, and
that would significantly reduce the probability in theory.

Signed-off-by: Xu Han <xuhan@redhat.com>