qemu_vm.py: Empty self.virtio_ports before fill it
'state' is not always None when initialize class VM, especially
when the test case is in non-first place of the test run. The list
will be superimposed in one loop since the parameter 'state' will
record it and reuse previous list. So need to be empty it before
fill it again

Signed-off-by: Qianqian Zhu <qizhu@redhat.com>