test_setup: provide option to check VF count on cleanup
currently `check_vfs_count()` does only check vf count if it
matches with hardcoded `driver_option` value during setup,
but it will be useful if we use same API to assert during
cleanup as well so provide an option to `check_vfs_count()`.

Also currently in `set_vf()` the return value of cmd that echos
the no of vfs in sysfs is checked, this commit would include
to check whether the vf count really matches after echo.

Signed-off-by: Balamuruhan S <bala24@linux.ibm.com>