setup: Improve the virtualenv check to work with python3's venv too
If one creates virtual environments using the python3's venv module and
then tries installing avocado VT into it, avocado tries to install into
the system path instead. This is because the detection of virtual
environment currently in place only works with the legacy virtualenv
module, not venv.
For venv, a different check needs to be in place according to:
https://docs.python.org/3/library/sys.html#sys.base_prefix

Signed-off-by: Erik Skultety <eskultet@redhat.com>