CI: add smoke checks based on Cirrus CI
This adds an initial set of Avocado + Avocado-VT basic set of
checks on Fedora 29 and 30.

Because Avocado-VT should work with LTS and non-LTS version, this uses
two released versions (one LTS, one regular release) and also adds a
compatibility check with Avocado's master.

This initial version is also limited to Python 2 (and given the
Python 2 lifetime, will probably remain like that).

Signed-off-by: Cleber Rosa <crosa@redhat.com>