virttest.qemu_devices: Allow cache='' in images definition
The newly introduced cache_map only works with valid values or None, but
these params are often specified in cart config and might result in
empty string instead of None. Even we set it to "" on line 2018 in
cdroms_define_by_params in case of aio == threads.

Instead of fixing our value let's treat any empty string the same as
None as users might want to get the default from cart config.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>