virttest.qemu_vm: Get _caps of the updated VM
During "make_create_command" the VM object is re-created with the new
params, which might affect the "self._caps" value, therefor we have to
use the new VM's _caps in this function.

This fixes a real issue where first test uses >2.12 qemu and second test
uses <2.12 qemu. In this case _caps are inherited from previous params,
therefor it attempts to use -blockdev.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>