CI: add smoke checks based on containers and Cirrus CI
This adds an initial set of Avocado + Avocado-VT basic set of
checks on:
  - latest Fedora (30)
  - CentOS 7
  - RHEL 8

Because Avocado-VT should work with Avocado's LTS and non-LTS version,
this uses the latest LTS release and Avocado's latest master.

The RHEL 8 checks are simpler, and only involve checking the presence
of the Avocado-VT plugins, instead of doing the full bootstrap, list
and dry-run, because of package limitations.  This could be improved
in the future with some moking of the required packages and binaries.

Signed-off-by: Cleber Rosa <crosa@redhat.com>