shared: Force anaconda output to serial console
With multiple console outputs Anaconda runs usually on the later and on
some archs not even there. Let's modify our "console=" lines to make it
always produce the output to serial console, which is captured in nice
format.

The only cons is that on ppc64 some messages will go only to serial
console and not to VGA output. This only affects some kernel messages
and not the graphical installation itself.

Note: We are already using this on Fedora and RHEL8 and it works better
than the previous implementation with output to tty0.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>