Refactor qemu capabilities related code
Currently, the capabilities of qemu is stored in `VM` and it would
result in some issue, for example, the inconsistency of this
attribute between `VM` and `DevContainer`.

To resolve the issue, as Lukáš suggested we can store the capabilities
only in `DevContainer` and make it to be a public attribute, hence,
it can be used by any context which needs that, and its lifetime will
be managed easily.

CC: Lukáš Doktor <ldoktor@redhat.com>
Signed-off-by: Xu Han <xuhan@redhat.com>