virttest.qemu_vm: Replace qemu_dst_binary on migrate
For years we had been claiming ping-pong migration, but in reality we do
ping->pong->pong->pong migrations as the source qemu_binary is only used
once. Let's replace the qemu_dst_binary for qemu_binary of the source VM
in order to start switching between ping and pong binaries.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>