virttest: Do not invalidate cache when raising error on get_address
Invalidating/dropping the cache is important when proceeding in the
testing, but it's not important when raising an exception and actually
breaks in case the VM is being migrated. Let's only raise the exception
while keeping the cache as sniffer should update in case of a change
anyway allowing the next _get_address() call to attempt to verify this
address again (should the user attempt to get the address for the second
time).

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>