virttest.qemu_vm: Properly cleanup src vm on migration
Previously we relied on a freshly-produced clone of the src vm to
cleanup the source vm on "finally" stage. This is not enough, because
the source vm (self) contains processes (nc attached to consoles). Let's
explicitly cleanup the source vm (self) just before replacing ourselves
with the destination (clone) vm to make sure all those processes are
cleared. To be super-safe let's keep the updated clone for the "finally"
postprocess.

Without this patch I'm getting about 2 extra aexpect (nc) processes per
ping_pong migration, with this patch it cleans everything properly.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>