qemu_vm: Avoid vnc ports clash on migration
The ports 590x are used for VNC (-vnc :0 => 5900) but the port of the
currently processed VM is not yet taken so find_free_port has no chance
on judging whether the port 5900 can be taken. This can happen with any
port depending on the current -vnc index so let's just limit the
available range to 5899 to completely avoid such clashes.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>