Fix vmx filename in '-i vmx' for v2v
The name of a guest in esx, the nfs path of esx guest files and the
name of vmx file may be different

For '-i vmx', It always needs to mount the nfs storage to local host,
so we can get the real vmx file name. But for '-i vmx -it ssh', NFS
storage don't have to be mounted, it's not worth mounting it and only
for getting the correct vmx file name. Luckily we can infer more than
95% guest's name by it's nfs disk path. And if the inference is wrong,
V2V will report an error. If v2v runs successfully, it can be sure
the name is correct.

If some corner guests cannot run successfully, you can set the correct
nfs source as in '-i vmx' for those corver scenarios.

Signed-off-by: Xiaodai Wang <xiaodwan@redhat.com>