cartesian_config: Introduce the new operator '~='
Introduce the new operator `~=` to support variable lazy assignment,
which means a variable will be set only if it has not been defined
in previous frames during configuration parsing. For example, by
given the following configuration, the value of the variable `foo`
should be `"bar"`,

    ```
    foo ~= bar
    ```

whereas, by given the following configuration, the value of the
variable `foo` would still be `"bar"`.

    ```
    foo = bar
    foo ~= foo
    ```

This might be useful for generating configuration under uncertain
conditions. For instance, we can set `smp` by using the value of
`vcpu_maxcpus` in the case of `smp` is not being set, otherwise,
let `smp` be what it was.

Signed-off-by: Xu Han <xuhan@redhat.com>