Remove usage of 52lts compatiblity layer by Avocado-VT itself
Because Avocado 52lts has not been supported for a while, we should
drop the compatibility layer.  This is a first stage, which removes
the *usage* of that layer by Avocado itself.

The test providers also need to drop the usage of the 52lts
compatibility functions, and then we can remove it from Avocado-VT
itself.

There's one exception category, which was not touched on this commit,
which is the staging libraries.  My hope is that we can determine the
fate of those before we have to change them.  It seems that the course
of action is to move pending functionality, if any, to Avocado itself,
and them drop those from staging.

This deserves some serious testing, as I don't expect to have gotten
all the conversions right for the very first time.

Signed-off-by: Cleber Rosa <crosa@redhat.com>