virttest.qemu_devices: Add default pcie root port on arm64-pci
Recently we changed the PCIE topology representation in virttest so it
uses multifunction and creates root ports for each added device, but
this topology requires the first root-port with multifunction=on plugged
to the first slot (0x0). It has already been added to q35 but aarch64
was left behind resulting in unbootable machine.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>