Add support for "-compat"
The "-compat" option was introduced recently in qemu to let users
configure the behavior on the use of deprecated interfaces. See:

    Configurable policy for handling deprecated interfaces

email for more details. Let's add this interface to both qemu_vm and
libvirt_vm to allow developing tests or simply running tests with custom
policy (the example in base.cfg makes the qemu crash when deprecated
interface is used).

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>