Fix libvirtd_log_cleanup for python3
For python3 the `libvirtd_debug_file` is always removed.
The reason is that there is a change in behavior which seems to
be intended for python3 as by https://bugs.python.org/issue35790:

In the `finally` block `sys.exc_info()` does not return the last
exception but the last unhandled exception (which is `(None, None,
None)` if only a single `exception.TestError` was raised, for example).

Store the `exc_info` in case of error and test that info which works for both
python2 and python3.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>