Use correct comparison between str/int and python literals
The compiler now produces a SyntaxWarning when identity checks
(is and is not) are used  with certain types of literals (e.g.
strings, numbers). These  can often work by  accident in CPython,
but are not  guaranteed by the language spec. The warning advises
users to use equality tests (== and !=) instead.

Reference: https://bugs.python.org/issue34850

Signed-off-by: Plamen Dimitrov <pdimitrov@pevogam.com>