Allow for errors while waiting
Some methods might raise, e.g. if a device is not ready yet.
Add new parameter in order to allow for ignoring errors.
Set new parameter default value in order to maintain legacy behavior.
Add pylint exception because in this specific case we want to handle all errors and exceptions.

Signed-off-by: Sebastian Mitterle <smitterl@redhat.com>