Requirements: update netaddr and zipp versions
Pinning netaddr 0.7.19 as the version for selftests (because 0.7.20
won't run on Python 2 / EL 7), and updating to that minimum version
for installation purposes (just to keep more consistency between
what's used to test and what's used to run it).

Also, the latest zipp gets pulled by other dependencies, and end up
failing to install properly on Python 2 / EL 7.

Signed-off-by: Cleber Rosa <crosa@redhat.com>