Remove usage of 52lts compatibility layer by Avocado-VT itself
Because Avocado 52lts has not been supported for a while, we should
drop the compatibility layer.  This is a first stage, which removes
the *usage* of that layer by Avocado itself.

The test providers also need to drop the usage of the 52lts
compatibility functions, and then we can remove it from Avocado-VT
itself.

Even the staging libraries were addressed, because it makes it allows
the compatibility module to be removed, and thus, prove that there are
no leftovers, and that there won't be any regressions.  The staging
libraries situation still need to be addressed in the sense that they
need to be synced or ideally moved to avocado.utils.

This still deserves some more testing, and that is really appreciated,
but I'm much more confident with it.  I've checked for obvious errors,
common on this type of mass change, with:

  inspekt lint --disable W,R,C,E1002,E1101,E1103,E1120,F0401,I0011,E1003

And there are no new issues introduced with this change.  I've also
checked and tp-qemu and tp-libvirt are already free of compat_52lts.

Signed-off-by: Cleber Rosa <crosa@redhat.com>