virttest/ceph.py: fix avocado.utils.process usage
On commit d9acd3a3a1 there was a mass change in favor of direct usage
of avocado.utils.process, instead of the now removed 52lts
compatibility code.  A few of those changes resulted in mistakes, such
as the issue reported by Yongxue Hong with `process.system_output()`
instead of `process.run()`.

Besides that one change, this defaults to using `process.run()`, and
while at it, fixes a bug that was using the result of
`process.system()` (the process return code) instead of the actual
output produced by the process.

Reported-by: Yongxue Hong <yhong@redhat.com>
Signed-off-by: Cleber Rosa <crosa@redhat.com>