libvirt check: delay the actual check until/if needed
The way the current check for the libvirt version is done, at the
module level, means that the code will called even if Avocado-VT
is never used or the check is never needed.  For instance, running
an Avocado job with a simple "passtest.py" results in:

   $ ./examples/jobs/passjob_html.py
   WARNING:root:libvirtd -V
   WARNING:root:libvirt version value by libvirtd or virtqemud command: ['libvirtd (libvirt) 5.6.0']
   JOB ID     : 1d7dfcb1e047acee1abc8cf656cb91f9f7463bd1
   JOB LOG    : /home/cleber/avocado/job-results/job-2020-07-06T09.57-1d7dfcb/job.log
    (1/1) examples/tests/passtest.py:PassTest.test: PASS (0.00 s)
   RESULTS    : PASS 1 | ERROR 0 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
   JOB HTML   : /home/cleber/avocado/job-results/job-2020-07-06T09.57-1d7dfcb/results.html

With this change, it will only be executed (once) if needed.

Signed-off-by: Cleber Rosa <crosa@redhat.com>