env_process: Fix wrong usage of getting cpu arch/family
The function get_cpu_arch() is now deprecated on master branch of
avocado, and the original usage of it is not outdated, which means
we could no longer get the info we want (e.g. power8 or power9)
by calling the function. The function that could provide the info
is 'get_family()' now, But there's no function 'get_family()'
on 69lts branch which we're currently using. This fix is a
compromised way to fit with both branches for us to get certain
cpu info.

Signed-off-by: Haijiao Zhao <haizhao@redhat.com>