Spell check: introduce the concept and command
Fixing all spelling issues within Avocado-VT is a huge task, and
I don't think we can solve it at once.  The goal of this patch
is to enable the command and an initial set of changes.

Anyone is then welcome to work on this once it's merged, fixing
issues and/or adding words to the ignore list.  At release time,
I can take an extra amount of effort to reduce the issues until
we declare it as a requirement like Avocado itself does.

Signed-off-by: Cleber Rosa <crosa@redhat.com>