virttest: Add migration_template.py
This module is used for cross feature testing with migration.
With this module, testing migration with specified feature can
be implemented easily. Before, when a specified feature is
automated, migration part is always left behind because of its
complexity. Now with this module, when we automate a specified
feature, we can implement the migration part by inheriting the
MigrationTemplate and copy&&paste the code related to specified
feature here.
Also updated base.cfg and migration.py accordingly.

Signed-off-by: Fangge Jin <fjin@redhat.com>