Avocado compatibility: implement Init plugin
This plugin will only be active when Avocado-VT is used with versions
of Avocado that support that interface, and support the new settings
interfaces.

It will register all options under the configuration file, so that a
configuration file will be respected on such environments, and default
values will also be present.

Signed-off-by: Cleber Rosa <crosa@redhat.com>