Remove the dependancy on utils_misc
This is to remove remote.py's dependant on utils_misc.
The exsiting dependancy is only for some utilities of log file.
Put the log file utilities into a new module to make it standalone.
Existing remote.py is too messed up, so we need to lessen the dependancy
as much as we can do.

Signed-off-by: Dan Zheng <dzheng@redhat.com>