VT Joblock plugin: be lenient with test suites containing Tasks
With the resolver/nrunner implementation, there won't be the
traditional "test factories" in the test suites, but Tasks.

This attempts to fetch the class from the test factory, and ignores if
it fails, which would happen if a Task is there instead.

This leniency isn't a problem because the joblock plugin should only
enforce a lock if a VT (VirtTest) test in within the test suite, which
won't happen at this point as there's no resolver for VT tests.

Fixes: https://github.com/avocado-framework/avocado/issues/4102
Signed-off-by: Cleber Rosa <crosa@redhat.com>