v2v: copy vddklib from NFS to local
If a server has bad network connection with NFS storage, v2v may
starts nbdkit timeout. This patch checks parts of files' md5sum in
vddklib and compare it with local vddklib, re-copy to local if they
are different. When cases use different vddklib versions, it checks
the local vddklib version and only copy it when the new vddklib does
not exist on local server.

All new vddklib are copied to local server in name 'vddklib_x'. The
'latest' is a link to currently using vddklib.

The directories look like:
/var/tmp/vddklib_dir/latest -> /var/tmp/vddk_libdir/vddklib_1
/var/tmp/vddklib_dir/vddklib_1
/var/tmp/vddklib_dir/vddklib_2
...

Signed-off-by: Xiaodai Wang <xiaodwan@redhat.com>