Use Python's stardard method for distributing data files
We've been forcing an "system" (aka RPM) approach with regard to the
location for data files, in Python code.  This is wrong, and the
Python code should use Python standard locations.

Optionally, the RPM packaging *may* tweak this standard use, and
replace the locations.

This solves the issue when Avocado-VT is installed via wheel files,
given that wheel files do not support the "data_files" given to
setuptools.

Fixes: https://github.com/avocado-framework/avocado-vt/issues/2193
Signed-off-by: Cleber Rosa <crosa@redhat.com>