Update customize_libvirt_config for modular daemons
There are 8 sub daemons instead of a libvirtd when modular daemons
are enabled. In previous code, it changed 'libvirtd' to 'virtqemud'
for modular daemons env which fixed some problems. But for now, we
still need to support other daemons. So it is extended to all
modular daemons.

In addition, the missing virtsecretd is added in this PR.

Signed-off-by: Yingshun Cui <yicui@redhat.com>