Guest listing: allow configuration of option names for filtering guests
The current hardcoded names of options names mean that, because of
the existing default for some option names such as 'vt.guest_os',
it will always filter by that name.

To be able to reuse this utility method and get a list of all guests
(or some), it's necessary to make that configurable.

Signed-off-by: Cleber Rosa <crosa@redhat.com>