Add encryption support for domain backup xml
With the implementation of libvirt RFE bz1811906, luks encryption
was supported during domain backup process. Now, the SCRATCH
of the pull-mode backup and the TARGET of the push-mode can be
encrypted with LUKS. And the usage is to add a <encrytion> subelement
under <scratch> or <target> elements. This commit is to support it.

Signed-off-by: Yi Sun <yisun@redhat.com>