Removal of handling of (legacy) deprecated entry point name
Many years ago, Avocado-VT (really virt-test) test entry points
would be named after the test name.  It was just another hurdle
that require people to mind the function name.

We're still checking for that, but I could not find any test
using the legacy naming, and it's about time we enforce the
new naming.

Signed-off-by: Cleber Rosa <crosa@redhat.com>