virt_vm: avoid of vm's unstable status
This function usually is invoked after vm starts. But sometimes on specific host, like
in openstack, the vm starts a little slow, so serial console login may get unexpected
result when recogizing the console output message. This is to make sure the vm is in
proper state before console message is prompted and recoginized correctly.

Signed-off-by: Dan Zheng <dzheng@redhat.com>