qdevices: Modify the virtiofsd daemon '-d' as an optional parameter
About virtio-fs, in the initially developed framework, we added the '-d' parameter
by default to start virtiofsd. This parameter will cause the 'virtiofs daemon'
to print detailed thread call information during the entire startup process.
But in recent tests, we found that this parameter will significantly reduce the
overall performance of virtio-fs. So I suggest to modify this default parameter
to become an optional parameter. If you need to open the debug mode for testing,
we can add 'fs_enable_debug_mode = yes' in tp-qemu.

Signed-off-by: Zhenyu Zhang <zhenyzha@redhat.com>