Fix qemu convert no space
Originally qemu-img convert command and umount command are run in one shell, which
causes the final ret code is zero even though qemu-img convert has failed with an
error, like 'no space left', because the ret code is overlaped by umount command which
is often zero. And when that error happens, the gluster volume is not cleaned up which
may cause /tmp folder is full on gluster server host/vm and no new volume can be created.
So this fix is to throw out the qemu-img convert error and clean up the volume before
throwing.

Signed-off-by: Dan Zheng <dzheng@redhat.com>