Compatibility with config files in early stages
In early stage of avocado run, there is a time when the vt options are
created, but they are not synchronized with values from config files.
This can affect some of the global variables which are set during this
time. This commit ensures the synchronization right after the options
are registered.

Reference: #2912
Signed-off-by: Jan Richter <jarichte@redhat.com>